WIP. Some semblance of working scrolling in the debugger

+++ src/Spice86/ViewModels/DebuggerLineViewModel.cs
@@ -20,10 +20,15 @@ public class DebuggerLineViewModel : ViewModelBase {
     public string ByteString { get; }
     public FunctionInformation? Function { get; }
     public SegmentedAddress SegmentedAddress { get; }
-    public long Address { get; }
+    
+    /// <summary>
+    /// The physical address of this instruction in memory.
+    /// </summary>
+    public uint Address { get; }
+    
     public bool ContinuesToNextInstruction => _info.FlowControl == FlowControl.Next;
     public bool CanBeSteppedOver => _info.FlowControl is FlowControl.Call or FlowControl.IndirectCall or FlowControl.Interrupt;
-    public long NextAddress => _info.NextIP32;
+    public uint NextAddress => _info.NextIP32;
     public string Disassembly => _info.ToString();
     public bool IsSelected { get; set; }
     public List<BreakpointViewModel> Breakpoints { get; }
@@ -58,7 +63,7 @@ public class DebuggerLineViewModel : ViewModelBase {
         
         // Log the current and new values
         Console.WriteLine($"DebuggerLine at {Address:X8}: IsCurrentInstruction changing from {_isCurrentInstruction} to {newValue} (CPU IP: {_cpuState.IpPhysicalAddress:X8})");
-        
+
         // Update the property
         IsCurrentInstruction = newValue;
     }
